refactor: 기존 Stock 도메인 DomesticStock으로 변경
